backend-overall/backend-mock/requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
fastapi
uvicorn
python-multipart
google-cloud-aiplatform 
google-cloud-texttospeech
langchain-google-vertexai
pymupdf